+++ linearmodels/iv/gmm.py
@@ -1,8 +1,6 @@
 """
 Covariance and weight estimation for GMM IV estimators
 """
-from __future__ import absolute_import, division, print_function
-
 from numpy import asarray, unique
 from numpy.linalg import inv